+++ molecool/tests/test_molecool.py
@@ -32,6 +32,15 @@ def test_calculate_angle():
 
     assert expected_angle == molecool.calculate_angle(r1, r2, r3, degrees=False)
 
+
+@pytest.mark.parametrize("r1, r2, r3, expected_angle", [
+    (np.array([1, 0, 0]), np.array([0, 0, 0]), np.array([0, 1, 0]), 90), 
+    (np.array([0, 0, -1]), np.array([0, 1, 0]), np.array([1, 0, 0]), 60)])
+def test_calculate_angle_many(r1, r2, r3, expected_angle):
+    """ testing that calculates the angle """
+ 
+    assert pytest.approx(expected_angle, abs=1e-5) == molecool.calculate_angle(r1, r2, r3, degrees=True)
+
 def test_build_bond_list():
     coordinates = np.array([
       [1, 1, 1], 
@@ -47,23 +56,3 @@ def test_build_bond_list():
     for bond_length in bonds.values():
         assert bond_length == 1.4
 
-
-def test_molecular_mass():
-    ''' function to test calculate_molecular_mass function  '''
-    symbols = ['C', 'H', 'H', 'H', 'H']
-
-    calculated_mass = molecool.calculate_molecular_mass(symbols)
-    actual_mass = 16.04
-
-    assert pytest.approx(actual_mass, abs=1e-2) == calculated_mass
-
-def test_center_of_mass():
-    symbols = np.array(['C', 'H', 'H', 'H', 'H'])
-    coordinates = np.array([[1,1,1], [2.4,1,1], [-0.4, 1, 1], [1, 1, 2.4], [1, 1, -0.4]])
-
-    center_of_mass = molecool.calculate_center_of_mass(symbols, coordinates)
-
-    expected_center = np.array([1.,1.,1.])
-
-    assert np.array_equal(center_of_mass, expected_center)
-
